Refactor routes fo retrieve data, and update commit hashes
