+++ webapp/env.js
@@ -1,8 +1,2 @@
 
 // Public environment variables available in the page, under the variable `env`
-
-module.exports = {
-  journals_data: 'https://raw.githubusercontent.com/greenelab/scihub/75ffe6e1422b0b6a9a0034f6e671dd34601327cd/data/journal-coverage.tsv',
-  publishers_data: 'https://raw.githubusercontent.com/greenelab/scihub/39ac7f29cc1a2d0465dcc36a07d3854242f95738/data/coverage-by-category.tsv',
-};
-
